refactor: drop redundant ERC20 ABI
